removed dynamic meta from post, linked post and detail pages to one component, made post forms dialogs
