+++ app/modules/components/CoreModules.tsx
@@ -33,7 +33,7 @@ export async function CoreModules() {
                         modules.required.map(module => (
                             <LinkBox
                                 key={module.id}
-                                title={`${module.id.toUpperCase()} / ${module.title}`}
+                                title={`${module.title} / ${module.id.toUpperCase()}`}
                                 href={`/modules/${module.id}`}
                                 className={"max-w-screen-sm flex-grow"}
                                 description={module.description || undefined}
